r37: fix -Wsign-compare in experimental/{padded_slice,reduction,slice_write,transformer}

+++ ttnn/cpp/ttnn/operations/experimental/reduction/fast_reduce_nc/device/fast_reduce_nc_device_operation.cpp
@@ -32,7 +32,7 @@ void FastReduceNCDeviceOperation::validate_on_program_cache_miss(
         (args.dim >= 0 && args.dim <= tt::tt_metal::MAX_NUM_DIMENSIONS - 2),
         "dim must be between 0 and {}.",
         tt::tt_metal::MAX_NUM_DIMENSIONS - 2);
-    TT_FATAL((args.dim < input_rank), "dim must be smaller than input tensor rank {}.", input_rank);
+    TT_FATAL((static_cast<size_t>(args.dim) < input_rank), "dim must be smaller than input tensor rank {}.", input_rank);
 }
 
 TensorSpec FastReduceNCDeviceOperation::compute_output_specs(
